initial commit
This is a port of the R code found at https://github.com/BioStatMatt/sas7bdat
This module can be used to read sas7bdat data files. It can also be used as a
standalone script to convert sas7bdat files to csv. Right now, 64 bit files and
compressed files are not supported.