Unicode fix for Python 3
Data that should be interpreted as a Date (by checking the format
against DATE_FORMAT_STRINGS can't succeed because the format is
read as type 'bytes' in Python 3, and all elements of
DATE_FORMAT_STRINGS are Unicode strings in Python 3. This
does the conversion of the column format attribute so the
comparison can succeed on Python 3.